送受信レートのカウンタが溢れるのを修正した
送受信レートのカウント時に一度に4MB以上のカウントをしようとすると桁溢れをおこしてしまっていたので、longにして溢れないようにした。

+++ PeerCastStation/PeerCastStation.Core/Utils.cs
@@ -126,7 +126,7 @@ namespace PeerCastStation.Core
 
   public class RateCounter
   {
-    private int count;
+    private long count;
     private int begin;
     private float rate;
     public int Duration { get; private set; }
